move deck_view into src/game/deckview
