+++ cockatrice/src/game/deckview/deck_view.cpp
@@ -1,15 +1,13 @@
 #include "deck_view.h"
 
-#include "../client/ui/theme_manager.h"
-#include "../game/cards/card_database.h"
-#include "../main.h"
-#include "../settings/cache_settings.h"
+#include "../../client/ui/theme_manager.h"
+#include "../../game/cards/card_database.h"
+#include "../../settings/cache_settings.h"
 #include "decklist.h"
 
 #include <QApplication>
 #include <QGraphicsSceneMouseEvent>
 #include <QMouseEvent>
-#include <QtMath>
 #include <algorithm>
 
 DeckViewCardDragItem::DeckViewCardDragItem(DeckViewCard *_item,
