+++ cockatrice/src/game/deckview/deck_view.h
@@ -1,7 +1,7 @@
 #ifndef DECKVIEW_H
 #define DECKVIEW_H
 
-#include "../game/cards/abstract_card_drag_item.h"
+#include "../../game/cards/abstract_card_drag_item.h"
 #include "pb/move_card_to_zone.pb.h"
 
 #include <QGraphicsScene>
